TEARDROP_MANAGER::UpdateTeardrops() Ensure created teardrop has a valid shape
A track can be connected to a pad when just crossing it. So we can create 2 teardrops,
one from pad to track start point and the other to track end point.
However this is acceptable only if the pad position is inside the track.
Otherwise the 2 teardrop shapes can be strange (and of course incorrect

Fixes https://gitlab.com/kicad/code/kicad/-/issues/21851