Simplify barcode implementation.
Remove support for point-editing of non-cardinally rotated
barcodes.

Remove support to separately update just the barcode symbol
or just the text.

Removes the point-editor behaviour of using the outside
bounding box.  Even when the barcode is inverted, it
sometimes feels like it should be growing/shrinking the
margin, not the barcode.  It's more predictable when the
corners are on the symbol.