TEARDROP_MANAGER::UpdateTeardrops() Ensure created teardrop has a valid shape V2
A track can be connected to a pad or a via  when just crossing or touching it.
So we can create 2 teardrops one from pad/via to track start point and the other
to track end point. However this is acceptable only if the pad/via position is inside
the track. Otherwise the 2 teardrop shapes can be strange (and of course incorrect)

Fixes https://gitlab.com/kicad/code/kicad/-/issues/21851