Add EnableUseAuiPerspective advanced option to allow use of aui LoadPerspective()
LoadPerspective allows a easy way to store/load frame geometry.
Unfortunately it easy create broken geometry frame (missing or misplaced toolbars
for instance) when the number or the type of items has changed between last saved
geometry and the geometry frame in use when loading old Perspective. And the only
one way to fix a broken geometry is to delete the config json files.

So the saved Perspective is loaded only if EnableUseAuiPerspective advanced option
is set to 1, mainly for developers.

Fixes https://gitlab.com/kicad/code/kicad/-/issues/21828