Router: Reduce greedyness of nearest ratsnest layer logic when placing vias
This was stopping layer swaps when placing vias if all the nearby
footprints / tracks were on the same layer as the currently active
layer. In this instance, assume the via is for another routing
reason (such as a hop over / under) and switch to the active
layer pair layer if it is valid for the active layer.