Fix false positive warning message when updating board from schematic.
The board connectivity algorithm creates pseudo net names for footprints
that have multiple pads with the same number using a "_N" suffix.  At
some point a change was made to the zone net properties dialog that
allowed choosing a pseudo net name as the net name assigned to the zone.
This caused the netlist updater to trigger a false "zone has no pads
connected" warning because the pseudo net name is not in the schematic
net list.

The warning message was not very helpful. Added the correct zone layers
and the net name assigned to the zone to the message for improved
clarity.

Fixes https://gitlab.com/kicad/code/kicad/-/issues/22133