Fix crash in footprint info cache sorting.
Apparently there is a code path that results in a null std::unique_ptr when
parsing the footprint info cache.  The fix is a rather crude check to see
if the pointer is null in the sort lambda.  This may not be the correct fix
but it fixes the crash when launching