ERC/DRC reports: list ignored checks
similar to the ERC/DRC dialogs, which list the tests not performed as
they are set to severity "ignore", that list of ignored checks is now
part of the text or JSON report as well. that is important to be able
to determine whether a particular violation would have been part of the
report if the whole class of error was not ignored.