Reset tools before loading new footprint in footprint chooser.
This may not actually fix this issue.  It is probably safer to reset all
tools before changing the footprint view just in case an unexpected event
occurs that causes a tool to attempt to access an object no longer valid
in the view.

Fixes https://gitlab.com/kicad/code/kicad/-/issues/21929