Prevent crash potential crash when sheet path objects are not valid.
Some how SCH_SHEET_PATH::Last() is returning a non-null pointer and yet
resulting in the sheet path being empty which when pop_back() is called
causes a crash.  This should not even be possible but has been observed
multiple times so add a check for an empty sheet path and avoid calling
pop_back().