IBIS: fix voltage polarity when driving pullup current sources.
Commit e50cb5c6fe swapped the voltage pins used to control the
power clamp and pullup elements in the SPICE model generated from
an IBIS model. That was based on figure 5.6 of the IBIS Modelling
Cookbook (v4), which shows that the voltage values in the power
clamp I-V tables are referenced to the supply voltage (Vcc) and
that positive voltage values represent voltages below the reference.
However, later on, figure 5.8 and table 5.6 show that although the
voltage values in the pullup I-V tables are similarly referenced to
the supply voltage, the voltage sense is reversed, with positive
voltage values there representing voltages above the reference. So
only the power clamp elements should have their voltage pins
swapped.